+++ sim/simulation.h
@@ -1,0 +1,8 @@
+#ifndef SIMULATION_H_
+#define SIMULATION_H_
+
+void GUI_Setup(void *pset, void *flush, int w, int h);
+
+void GUI_Process();
+
+#endif